requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0

# Scheduling
APScheduler==3.10.4

# Template Engine
jinja2==3.1.2

# Environment
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1